fix: handle semver outputs independently for major and semver releases
Allow major release step to run if nextMajor is set, even if next
is not set. This ensures that major version tags (like v2) can be
updated even if the semver-action doesn't set the next output.

Also add a check to skip the major release step if the tag is empty.